requirements.txt
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.1.8
Flask==3.1.3
idna==3.11
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
psycopg2-binary==2.9.11
pygame==2.6.1
requests==2.32.5
urllib3==2.6.3
Werkzeug==3.1.8
zipp==3.23.1
